requirements.txt
python-dotenv
boto3
requests_aws4auth
opensearch-py